- job file 삭제 config 설정 - 검증 정보 전달 rest 생성

+++ src/main/java/com/seculayer/mrms/checker/DACompleteChecker.java
@@ -37,13 +37,15 @@ public class DACompleteChecker extends Checker {
     }
 
     private void removeInitFolder(String datasetID) {
-        String folderPath = MRMServerManager.getInstance().getConfiguration().get("ape.job.dir") + "/" + datasetID;
-        File f = new File(folderPath);
-        try {
-            FileUtils.cleanDirectory(f);
-            f.delete();
-        } catch (Exception e) {
-            e.printStackTrace();
+        if (MRMServerManager.getInstance().getConfiguration().getBoolean("jobfile.remove.boolean", false)) {
+            String folderPath = MRMServerManager.getInstance().getConfiguration().get("ape.job.dir") + "/" + datasetID;
+            File f = new File(folderPath);
+            try {
+                FileUtils.cleanDirectory(f);
+                f.delete();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
     }
